+++ dongamboard/app/src/main/java/com/study/dongamboard/activity/notice/NoticeCreateActivity.kt
@@ -1,10 +1,11 @@
-package com.study.dongamboard
+package com.study.dongamboard.activity.notice
 
 import android.os.Bundle
 import android.util.Log
 import android.widget.EditText
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
+import com.study.dongamboard.R
 import com.study.dongamboard.db.NoticeDB
 import com.study.dongamboard.model.NoticeData
 
@@ -34,12 +35,10 @@ class NoticeCreateActivity : AppCompatActivity() {
             createNoticeThread.start()
             finish()
         }
-
     }
 
     override fun onDestroy() {
         NoticeDB.destroyInstance()
         super.onDestroy()
     }
-
 }
