+++ dongamboard/app/src/main/java/com/study/dongamboard/activity/post/PostActivity.kt
@@ -1,4 +1,4 @@
-package com.study.dongamboard
+package com.study.dongamboard.activity.post
 
 import android.content.Context
 import android.content.DialogInterface
