+++ dongamboard/app/src/main/java/com/study/dongamboard/activity/post/PostUpdateActivity.kt
@@ -1,42 +1,45 @@
-package com.study.dongamboard
+package com.study.dongamboard.activity.post
 
 import android.os.Bundle
 import android.widget.EditText
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
+import com.study.dongamboard.R
 import com.study.dongamboard.api.APIObject
 import com.study.dongamboard.db.PostDB
+import com.study.dongamboard.model.response.PostResponse
 import com.study.dongamboard.model.request.PostRequest
-import com.study.dongamboard.type.BoardCategoryType
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class PostCreateActivity : AppCompatActivity() {
+class PostUpdateActivity : AppCompatActivity() {
 
-    lateinit var category: BoardCategoryType
-    lateinit var postDB : PostDB
+    lateinit var post: PostResponse
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_write)
 
-        category = intent.getSerializableExtra("postCategory") as BoardCategoryType
-        postDB = PostDB.getInstance(this)!!
+        post = intent.getSerializableExtra("postData") as PostResponse
 
         val etPostCreateTitle = findViewById<EditText>(R.id.etPostCreateTitle)
         val etPostCreateContent = findViewById<EditText>(R.id.etPostCreateContent)
 
+        etPostCreateTitle.setText(post.title)
+        etPostCreateContent.setText(post.content)
+
         val ivCreatePostBtn = findViewById<ImageView>(R.id.ivCreatePostBtn)
-        ivCreatePostBtn.setOnClickListener{
+        ivCreatePostBtn.setOnClickListener {
             CoroutineScope(Dispatchers.IO).launch {
                 val postRequest = PostRequest(
                     etPostCreateTitle.text.toString(),
                     etPostCreateContent.text.toString(),
-                    category)
-                APIObject.getRetrofitAPIService.createPost(postRequest)
+                    post.category!!
+                )
+                APIObject.getRetrofitAPIService.updatePost(post.id, postRequest)
+                finish()
             }
-            finish()
         }
     }
